# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: CivicX API
  version: "1.0.0"
  description: |
    CivicX — Civic engagement platform API.
    Shared contract between web (React), mobile (Flutter) and backend (Node/Express/TypeScript).
  contact:
    name: CivicX Backend Team

servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.civicx.com
    description: Production (future)

# ──────────────────────────────────────────────
# Reusable components
# ──────────────────────────────────────────────
components:
  schemas:
    # ── Enums ──────────────────────────────────
    Role:
      type: string
      enum:
        - citizen
        - university
        - industry
        - admin
      description: User roles across the platform.

    IssueStatus:
      type: string
      enum:
        - draft
        - open
        - under_review
        - in_progress
        - resolved
        - closed
        - rejected
      description: Lifecycle statuses for a civic issue.

    IssuePriority:
      type: string
      enum:
        - low
        - medium
        - high
        - critical

    IssueCategory:
      type: string
      enum:
        - infrastructure
        - safety
        - environment
        - transportation
        - community
        - education
        - health
        - governance
        - other

    # ── Standard envelopes ─────────────────────
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          description: Response payload (varies by endpoint).
        meta:
          type: object
          properties:
            requestId:
              type: string
              format: uuid
            timestamp:
              type: string
              format: date-time

    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error message.
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          type: object
          properties:
            requestId:
              type: string
              format: uuid

    # ── Health ─────────────────────────────────
    HealthData:
      type: object
      required:
        - status
        - timestamp
        - uptime
        - environment
      properties:
        status:
          type: string
          enum:
            - healthy
            - degraded
            - unhealthy
        timestamp:
          type: string
          format: date-time
        uptime:
          type: number
          description: Server uptime in seconds.
        environment:
          type: string
          enum:
            - development
            - staging
            - production
        version:
          type: string
        mongo:
          type: string
          enum:
            - connected
            - disconnected

    AuthUser:
      type: object
      required: [id, name, email, role, accountStatus, language]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        accountStatus: { type: string, enum: [pending, active, suspended] }
        institutionId: { type: [string, "null"] }
        phone: { type: string }
        language: { type: string, enum: [en, hi] }

    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 128 }
        phone: { type: string, maxLength: 30 }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 1, maxLength: 128 }

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 32, maxLength: 512 }

    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }

    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token: { type: string, minLength: 32, maxLength: 512 }
        password: { type: string, minLength: 8, maxLength: 128 }

    OnboardingRequest:
      type: object
      required: [name, email, password, institutionName, institutionType, domains, registrationNumber, contactPhone]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 128 }
        institutionName: { type: string, minLength: 2, maxLength: 200 }
        institutionType: { type: string, enum: [university, industry] }
        domains: { type: array, minItems: 1, maxItems: 20, items: { type: string, minLength: 2, maxLength: 80 } }
        registrationNumber: { type: string, minLength: 2, maxLength: 100 }
        contactPhone: { type: string, minLength: 5, maxLength: 30 }

  # ── Shared parameters ──────────────────────
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Client-supplied or server-generated request correlation ID.

  # ── Shared responses ──────────────────────
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: "Request validation failed"
              details:
                - field: "title"
                  message: "Title is required"
            meta:
              requestId: "550e8400-e29b-41d4-a716-446655440000"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: NOT_FOUND
              message: "The requested resource was not found"
            meta:
              requestId: "550e8400-e29b-41d4-a716-446655440000"

    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: "Authentication is required"

    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: "An unexpected error occurred"

# ──────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────
paths:
  /api/auth/register:
    post:
      operationId: registerCitizen
      summary: Register a citizen account
      tags: [Authentication]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RegisterRequest" } } }
      responses:
        "201": { description: Citizen registered }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Email already registered }
  /api/auth/onboard-request:
    post:
      operationId: createOnboardingRequest
      summary: Submit an institution onboarding request
      tags: [Authentication]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/OnboardingRequest" } } }
      responses:
        "201": { description: Pending institution request created }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Email already registered }
  /api/auth/web/login:
    post:
      operationId: webLogin
      summary: Sign in a web client
      tags: [Authentication]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LoginRequest" } } }
      responses:
        "200": { description: Access token returned and refresh cookie set }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Account pending approval or suspended }
  /api/auth/web/refresh:
    post:
      operationId: webRefresh
      summary: Rotate the web refresh cookie
      tags: [Authentication]
      responses:
        "200": { description: Rotated access token and refresh cookie }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/web/logout:
    post:
      operationId: webLogout
      summary: Revoke the web refresh cookie
      tags: [Authentication]
      responses:
        "204": { description: Logged out }
  /api/auth/mobile/login:
    post:
      operationId: mobileLogin
      summary: Sign in a mobile client
      tags: [Authentication]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LoginRequest" } } }
      responses:
        "200": { description: Access and refresh tokens returned }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Account pending approval or suspended }
  /api/auth/mobile/refresh:
    post:
      operationId: mobileRefresh
      summary: Rotate a mobile refresh token
      tags: [Authentication]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RefreshRequest" } } }
      responses:
        "200": { description: Rotated access and refresh tokens }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/mobile/logout:
    post:
      operationId: mobileLogout
      summary: Revoke a mobile refresh token
      tags: [Authentication]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RefreshRequest" } } }
      responses:
        "204": { description: Logged out }
  /api/auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the authenticated user
      tags: [Authentication]
      responses:
        "200": { description: Current user }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Request a password-reset token
      tags: [Authentication]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ForgotPasswordRequest" } } }
      responses:
        "202": { description: Generic response whether or not the email exists }
  /api/auth/reset-password:
    post:
      operationId: resetPassword
      summary: Reset a password using an unexpired token
      tags: [Authentication]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ResetPasswordRequest" } } }
      responses:
        "204": { description: Password reset and refresh sessions revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/onboarding/{id}/approve:
    patch:
      operationId: approveOnboarding
      summary: Approve a pending institution request
      tags: [Authentication]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Institution and its pending user are activated }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Admin role required }
  /api/health:
    get:
      operationId: getHealth
      summary: Health check
      description: Returns backend health status. Used by all clients and monitoring.
      tags:
        - Health
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Backend is healthy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HealthData"
              example:
                success: true
                data:
                  status: healthy
                  timestamp: "2025-09-06T10:30:00.000Z"
                  uptime: 3600
                  environment: development
                  version: "1.0.0"
                  mongo: connected
                meta:
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
                  timestamp: "2025-09-06T10:30:00.000Z"
        "503":
          description: Backend is unhealthy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HealthData"

tags:
  - name: Authentication
    description: Registration, onboarding, sessions, and password recovery.
  - name: Health
    description: System health and readiness endpoints