# server/backend/src/modules/auth/auth.controller.ts
import type { CookieOptions, NextFunction, Request, Response } from "express";
import { Types } from "mongoose";
import config from "../../config";
import { ConflictError, ForbiddenError, NotFoundError, UnauthorizedError } from "../../utils/errors";
import { sendSuccess } from "../../utils/response";
import { Institution } from "./institution.model";
import { PasswordReset } from "./password-reset.model";
import { User, type UserDocument } from "./user.model";
import { createRefreshSession, hashPassword, newOpaqueToken, normalizeEmail, publicUser, revokeAllRefreshSessions, revokeRefreshSession, rotateRefreshSession, signAccessToken, tokenHash, verifyPassword } from "./auth.service";
import type { ClientType } from "./auth.types";

const cookieOptions: CookieOptions = { httpOnly: true, sameSite: "lax", secure: config.isProd, path: "/api/auth/web" };
const readCookie = (req: Request, name: string): string | undefined => req.headers.cookie?.split(";").map((part) => part.trim()).find((part) => part.startsWith(`${name}=`))?.slice(name.length + 1);
async function issue(user: UserDocument, client: ClientType) { const accessToken = await signAccessToken(user); const refresh = await createRefreshSession(user._id, client); return { accessToken, refreshToken: refresh.token, expiresAt: refresh.expiresAt }; }
function sendCredentials(res: Response, user: UserDocument, tokens: { accessToken: string; refreshToken: string; expiresAt: Date }, client: ClientType): void { if (client === "web") { res.cookie("refreshToken", tokens.refreshToken, { ...cookieOptions, expires: tokens.expiresAt }); sendSuccess(res, { user: publicUser(user), accessToken: tokens.accessToken }); return; } sendSuccess(res, { user: publicUser(user), accessToken: tokens.accessToken, refreshToken: tokens.refreshToken, refreshExpiresAt: tokens.expiresAt }); }

export async function register(req: Request, res: Response, next: NextFunction): Promise<void> { try { const { name, email, password } = req.body as { name: string; email: string; password: string }; const normalized = normalizeEmail(email); if (await User.exists({ email: normalized })) throw ConflictError("An account with this email already exists"); const user = await User.create({ name, email: normalized, passwordHash: await hashPassword(password), role: "citizen", accountStatus: "active" }); sendSuccess(res, { user: publicUser(user) }, 201); } catch (error) { next(error); } }
export async function onboard(req: Request, res: Response, next: NextFunction): Promise<void> { try { const body = req.body as { name: string; email: string; password: string; institutionName: string; institutionType: "university" | "industry" }; const email = normalizeEmail(body.email); if (await User.exists({ email })) throw ConflictError("An account with this email already exists"); const institution = await Institution.create({ name: body.institutionName, type: body.institutionType, accountStatus: "pending" }); const user = await User.create({ name: body.name, email, passwordHash: await hashPassword(body.password), role: body.institutionType, accountStatus: "pending", institutionId: institution._id }); sendSuccess(res, { requestId: institution._id.toString(), status: "pending", user: publicUser(user) }, 201); } catch (error) { next(error); } }
async function login(req: Request, res: Response, next: NextFunction, client: ClientType): Promise<void> { try { const { email, password } = req.body as { email: string; password: string }; const user = await User.findOne({ email: normalizeEmail(email) }).select("+passwordHash"); if (!user || !(await verifyPassword(user.passwordHash, password))) throw UnauthorizedError("Invalid email or password"); if (user.accountStatus !== "active") throw ForbiddenError(user.accountStatus === "pending" ? "This institution account is pending approval" : "This account has been suspended"); sendCredentials(res, user, await issue(user, client), client); } catch (error) { next(error); } }
export const webLogin = (req: Request, res: Response, next: NextFunction) => login(req, res, next, "web");
export const mobileLogin = (req: Request, res: Response, next: NextFunction) => login(req, res, next, "mobile");
async function refresh(req: Request, res: Response, next: NextFunction, client: ClientType): Promise<void> { try { const token = client === "web" ? readCookie(req, "refreshToken") : (req.body as { refreshToken: string }).refreshToken; if (!token) throw UnauthorizedError("Refresh token is required"); const rotated = await rotateRefreshSession(token, client); sendCredentials(res, rotated.user, { accessToken: await signAccessToken(rotated.user), refreshToken: rotated.refreshToken, expiresAt: rotated.expiresAt }, client); } catch { next(UnauthorizedError("Invalid or expired refresh token")); } }
export const webRefresh = (req: Request, res: Response, next: NextFunction) => refresh(req, res, next, "web");
export const mobileRefresh = (req: Request, res: Response, next: NextFunction) => refresh(req, res, next, "mobile");
async function logout(req: Request, res: Response, next: NextFunction, client: ClientType): Promise<void> { try { const token = client === "web" ? readCookie(req, "refreshToken") : (req.body as { refreshToken?: string }).refreshToken; if (token) await revokeRefreshSession(token); if (client === "web") res.clearCookie("refreshToken", cookieOptions); res.status(204).send(); } catch (error) { next(error); } }
export const webLogout = (req: Request, res: Response, next: NextFunction) => logout(req, res, next, "web");
export const mobileLogout = (req: Request, res: Response, next: NextFunction) => logout(req, res, next, "mobile");
export async function me(req: Request, res: Response, next: NextFunction): Promise<void> { try { const user = await User.findById(req.auth?.userId); if (!user) throw UnauthorizedError(); sendSuccess(res, { user: publicUser(user) }); } catch (error) { next(error); } }
export async function forgotPassword(req: Request, res: Response, next: NextFunction): Promise<void> { try { const user = await User.findOne({ email: normalizeEmail((req.body as { email: string }).email) }); if (user?.accountStatus === "active") { const token = newOpaqueToken(); await PasswordReset.updateMany({ userId: user._id, usedAt: { $exists: false } }, { $set: { usedAt: new Date() } }); await PasswordReset.create({ userId: user._id, tokenHash: tokenHash(token), expiresAt: new Date(Date.now() + config.passwordResetTtlMinutes * 60000) }); } sendSuccess(res, { message: "If an account exists, password reset instructions have been sent." }, 202); } catch (error) { next(error); } }
export async function resetPassword(req: Request, res: Response, next: NextFunction): Promise<void> { try { const { token, password } = req.body as { token: string; password: string }; const reset = await PasswordReset.findOne({ tokenHash: tokenHash(token), usedAt: { $exists: false }, expiresAt: { $gt: new Date() } }); if (!reset) throw UnauthorizedError("Invalid or expired reset token"); const user = await User.findById(reset.userId).select("+passwordHash"); if (!user) throw UnauthorizedError("Invalid or expired reset token"); user.passwordHash = await hashPassword(password); reset.usedAt = new Date(); await Promise.all([user.save(), reset.save(), revokeAllRefreshSessions(user._id)]); res.status(204).send(); } catch (error) { next(error); } }
export async function approveOnboarding(req: Request, res: Response, next: NextFunction): Promise<void> { try { const id = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id; if (!id || !Types.ObjectId.isValid(id)) throw NotFoundError("Institution request not found"); const institution = await Institution.findById(id); if (!institution) throw NotFoundError("Institution request not found"); institution.accountStatus = "active"; await Promise.all([institution.save(), User.updateMany({ institutionId: institution._id, accountStatus: "pending" }, { $set: { accountStatus: "active" } })]); sendSuccess(res, { requestId: institution._id.toString(), status: "approved" }); } catch (error) { next(error); } }